+++ src/main/java/kitchenpos/order/domain/JdbcTemplateOrderDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.model.Order;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
