+++ src/main/java/kitchenpos/order/domain/OrderDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.model.Order;
+import kitchenpos.order.domain.Order;
 
 import java.util.List;
 import java.util.Optional;
